Refactor ETL runner into modular package
